PaymentManager introduced and payment-related loops factored out from Pool
