+++ pool/pool.py
@@ -4,8 +4,7 @@ import pathlib
 import time
 import traceback
 from asyncio import Task
-from math import floor
-from typing import Dict, Optional, Set, List, Tuple, Callable
+from typing import Dict, Optional, Tuple, Callable
 
 import os
 import yaml
@@ -23,7 +22,6 @@ from chia.protocols.pool_protocol import (
     POOL_PROTOCOL_VERSION,
 )
 from chia.rpc.wallet_rpc_client import WalletRpcClient
-from chia.types.blockchain_format.coin import Coin
 from chia.types.coin_record import CoinRecord
 from chia.types.coin_solution import CoinSolution
 from chia.util.bech32m import decode_puzzle_hash
@@ -38,16 +36,16 @@ from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.consensus.pot_iterations import calculate_iterations_quality
 from chia.util.lru_cache import LRUCache
 from chia.util.chia_logging import initialize_logging
-from chia.wallet.transaction_record import TransactionRecord
 from chia.pools.pool_puzzles import (
-    get_most_recent_singleton_coin_from_coin_solution,
     get_delayed_puz_info_from_launcher_spend,
     launcher_id_to_p2_puzzle_hash,
 )
 
 from .blockchain_state import StateKeeper
 from .difficulty_adjustment import get_new_difficulty
-from .singleton import create_absorb_transaction, get_singleton_state, get_coin_spend
+from .payment_manger.abstract import AbstractPaymentManager
+from .payment_manger.default import DefaultPaymentManager
+from .singleton import get_singleton_state, get_coin_spend
 from .store.abstract import AbstractPoolStore
 from .store.sqlite_store import SqlitePoolStore
 from .record import FarmerRecord
@@ -56,7 +54,8 @@ from .util import error_dict
 
 class Pool:
     def __init__(self, config: Dict, constants: ConsensusConstants, pool_store: Optional[AbstractPoolStore] = None,
-                 difficulty_function: Callable = get_new_difficulty):
+                 difficulty_function: Callable = get_new_difficulty,
+                 payment_manger: Optional[DefaultPaymentManager] = None):
         self.follow_singleton_tasks: Dict[bytes32, asyncio.Task] = {}
         self.log = logging
         # If you want to log to a file: use filename='example.log', encoding='utf-8'
@@ -79,8 +78,8 @@ class Pool:
         self.constants = constants
 
         self.store: AbstractPoolStore = pool_store or SqlitePoolStore()
-
-        self.pool_fee = pool_config["pool_fee"]
+        self.payment_manager: AbstractPaymentManager = payment_manger or DefaultPaymentManager(
+            self.log, pool_config, self.constants)
 
         # This number should be held constant and be consistent for every pool in the network. DO NOT CHANGE
         self.iters_limit = self.constants.POOL_SUB_SLOT_ITERS // 64
@@ -110,10 +109,6 @@ class Pool:
         # Using 2164248527
         self.default_target_puzzle_hash: bytes32 = bytes32(decode_puzzle_hash(pool_config["default_target_address"]))
 
-        # The pool fees will be sent to this address. This MUST be on a different key than the target_puzzle_hash,
-        # otherwise, the fees will be sent to the users. Using 690783650
-        self.pool_fee_puzzle_hash: bytes32 = bytes32(decode_puzzle_hash(pool_config["pool_fee_address"]))
-
         # This is the wallet fingerprint and ID for the wallet spending the funds from `self.default_target_puzzle_hash`
         self.wallet_fingerprint = pool_config["wallet_fingerprint"]
         self.wallet_id = pool_config["wallet_id"]
@@ -131,27 +126,8 @@ class Pool:
         self.farmer_update_blocked: set = set()
         self.farmer_update_cooldown_seconds: int = 600
 
-        # These are the phs that we want to look for on chain, that we can claim to our pool
-        self.scan_p2_singleton_puzzle_hashes: Set[bytes32] = set()
-
-        # Don't scan anything before this height, for efficiency (for example pool start date)
-        self.scan_start_height: uint32 = uint32(pool_config["scan_start_height"])
-
-        # Interval for scanning and collecting the pool rewards
-        self.collect_pool_rewards_interval = pool_config["collect_pool_rewards_interval"]
-
         # After this many confirmations, a transaction is considered final and irreversible
         self.confirmation_security_threshold = pool_config["confirmation_security_threshold"]
-
-        # Interval for making payout transactions to farmers
-        self.payment_interval = pool_config["payment_interval"]
-
-        # We will not make transactions with more targets than this, to ensure our transaction gets into the blockchain
-        # faster.
-        self.max_additions_per_transaction = pool_config["max_additions_per_transaction"]
-
-        # This is the list of payments that we have not sent yet, to farmers
-        self.pending_payments: Optional[asyncio.Queue] = None
 
         # We target these many partials for this number of seconds. We adjust after receiving this many partials.
         self.number_of_partials_target: int = pool_config["number_of_partials_target"]
@@ -159,9 +135,6 @@ class Pool:
 
         # Tasks (infinite While loops) for different purposes
         self.confirm_partials_loop_task: Optional[asyncio.Task] = None
-        self.collect_pool_rewards_loop_task: Optional[asyncio.Task] = None
-        self.create_payment_loop_task: Optional[asyncio.Task] = None
-        self.submit_payment_loop_task: Optional[asyncio.Task] = None
 
         self.node_rpc_client: Optional[FullNodeRpcClient] = None
         self.node_rpc_port = pool_config["node_rpc_port"]
@@ -178,15 +151,9 @@ class Pool:
         await self.init_wallet_rpc()
 
         await self.state_keeper.start(self.node_rpc_client, self.wallet_rpc_client)
-
-        self.scan_p2_singleton_puzzle_hashes = await self.store.get_pay_to_singleton_phs()
+        await self.payment_manager.start(self.node_rpc_client, self.wallet_rpc_client, self.store, self.state_keeper)
 
         self.confirm_partials_loop_task = asyncio.create_task(self.confirm_partials_loop())
-        self.collect_pool_rewards_loop_task = asyncio.create_task(self.collect_pool_rewards_loop())
-        self.create_payment_loop_task = asyncio.create_task(self.create_payment_loop())
-        self.submit_payment_loop_task = asyncio.create_task(self.submit_payment_loop())
-
-        self.pending_payments = asyncio.Queue()
 
     async def init_node_rpc(self):
         self.node_rpc_client = await FullNodeRpcClient.create(
@@ -207,12 +174,6 @@ class Pool:
     async def stop(self):
         if self.confirm_partials_loop_task is not None:
             self.confirm_partials_loop_task.cancel()
-        if self.collect_pool_rewards_loop_task is not None:
-            self.collect_pool_rewards_loop_task.cancel()
-        if self.create_payment_loop_task is not None:
-            self.create_payment_loop_task.cancel()
-        if self.submit_payment_loop_task is not None:
-            self.submit_payment_loop_task.cancel()
 
         await self.state_keeper.stop()
         self.wallet_rpc_client.close()
@@ -220,243 +181,6 @@ class Pool:
         self.node_rpc_client.close()
         await self.node_rpc_client.await_closed()
         await self.store.close()
-
-    async def collect_pool_rewards_loop(self):
-        """
-        Iterates through the blockchain, looking for pool rewards, and claims them, creating a transaction to the
-        pool's puzzle_hash.
-        """
-
-        while True:
-            try:
-                if not self.state_keeper.blockchain_state["sync"]["synced"]:
-                    await asyncio.sleep(60)
-                    continue
-
-                scan_phs: List[bytes32] = list(self.scan_p2_singleton_puzzle_hashes)
-                peak_height = self.state_keeper.blockchain_state["peak"].height
-
-                # Only get puzzle hashes with a certain number of confirmations or more, to avoid reorg issues
-                coin_records: List[CoinRecord] = await self.node_rpc_client.get_coin_records_by_puzzle_hashes(
-                    scan_phs,
-                    include_spent_coins=False,
-                    start_height=self.scan_start_height,
-                )
-                self.log.info(
-                    f"Scanning for block rewards from {self.scan_start_height} to {peak_height}. "
-                    f"Found: {len(coin_records)}"
-                )
-                ph_to_amounts: Dict[bytes32, int] = {}
-                ph_to_coins: Dict[bytes32, List[CoinRecord]] = {}
-                not_buried_amounts = 0
-                for cr in coin_records:
-                    if cr.confirmed_block_index > peak_height - self.confirmation_security_threshold:
-                        not_buried_amounts += cr.coin.amount
-                        continue
-                    if cr.coin.puzzle_hash not in ph_to_amounts:
-                        ph_to_amounts[cr.coin.puzzle_hash] = 0
-                        ph_to_coins[cr.coin.puzzle_hash] = []
-                    ph_to_amounts[cr.coin.puzzle_hash] += cr.coin.amount
-                    ph_to_coins[cr.coin.puzzle_hash].append(cr)
-
-                # For each p2sph, get the FarmerRecords
-                farmer_records = await self.store.get_farmer_records_for_p2_singleton_phs(
-                    set([ph for ph in ph_to_amounts.keys()])
-                )
-
-                # For each singleton, create, submit, and save a claim transaction
-                claimable_amounts = 0
-                not_claimable_amounts = 0
-                for rec in farmer_records:
-                    if rec.is_pool_member:
-                        claimable_amounts += ph_to_amounts[rec.p2_singleton_puzzle_hash]
-                    else:
-                        not_claimable_amounts += ph_to_amounts[rec.p2_singleton_puzzle_hash]
-
-                if len(coin_records) > 0:
-                    self.log.info(f"Claimable amount: {claimable_amounts / (10**12)}")
-                    self.log.info(f"Not claimable amount: {not_claimable_amounts / (10**12)}")
-                    self.log.info(f"Not buried amounts: {not_buried_amounts / (10**12)}")
-
-                for rec in farmer_records:
-                    if rec.is_pool_member:
-                        singleton_tip: Optional[Coin] = get_most_recent_singleton_coin_from_coin_solution(
-                            rec.singleton_tip
-                        )
-                        if singleton_tip is None:
-                            continue
-
-                        singleton_coin_record: Optional[
-                            CoinRecord
-                        ] = await self.node_rpc_client.get_coin_record_by_name(singleton_tip.name())
-                        if singleton_coin_record is None:
-                            continue
-                        if singleton_coin_record.spent:
-                            self.log.warning(
-                                f"Singleton coin {singleton_coin_record.coin.name()} is spent, will not "
-                                f"claim rewards"
-                            )
-                            continue
-
-                        spend_bundle = await create_absorb_transaction(
-                            self.node_rpc_client,
-                            rec,
-                            self.state_keeper.blockchain_state["peak"].height,
-                            ph_to_coins[rec.p2_singleton_puzzle_hash],
-                            self.constants.GENESIS_CHALLENGE,
-                        )
-
-                        if spend_bundle is None:
-                            continue
-
-                        push_tx_response: Dict = await self.node_rpc_client.push_tx(spend_bundle)
-                        if push_tx_response["status"] == "SUCCESS":
-                            # TODO(pool): save transaction in records
-                            self.log.info(f"Submitted transaction successfully: {spend_bundle.name().hex()}")
-                        else:
-                            self.log.error(f"Error submitting transaction: {push_tx_response}")
-                await asyncio.sleep(self.collect_pool_rewards_interval)
-            except asyncio.CancelledError:
-                self.log.info("Cancelled collect_pool_rewards_loop, closing")
-                return
-            except Exception as e:
-                error_stack = traceback.format_exc()
-                self.log.error(f"Unexpected error in collect_pool_rewards_loop: {e} {error_stack}")
-                await asyncio.sleep(self.collect_pool_rewards_interval)
-
-    async def create_payment_loop(self):
-        """
-        Calculates the points of each farmer, and splits the total funds received into coins for each farmer.
-        Saves the transactions that we should make, to `amount_to_distribute`.
-        """
-        while True:
-            try:
-                if not self.state_keeper.blockchain_state["sync"]["synced"]:
-                    self.log.warning("Not synced, waiting")
-                    await asyncio.sleep(60)
-                    continue
-
-                if self.pending_payments.qsize() != 0:
-                    self.log.warning(f"Pending payments ({self.pending_payments.qsize()}), waiting")
-                    await asyncio.sleep(60)
-                    continue
-
-                self.log.info("Starting to create payment")
-
-                coin_records: List[CoinRecord] = await self.node_rpc_client.get_coin_records_by_puzzle_hash(
-                    self.default_target_puzzle_hash, include_spent_coins=False
-                )
-
-                if len(coin_records) == 0:
-                    self.log.info("No funds to distribute.")
-                    await asyncio.sleep(120)
-                    continue
-
-                total_amount_claimed = sum([c.coin.amount for c in coin_records])
-                pool_coin_amount = int(total_amount_claimed * self.pool_fee)
-                amount_to_distribute = total_amount_claimed - pool_coin_amount
-
-                self.log.info(f"Total amount claimed: {total_amount_claimed / (10 ** 12)}")
-                self.log.info(f"Pool coin amount (includes blockchain fee) {pool_coin_amount  / (10 ** 12)}")
-                self.log.info(f"Total amount to distribute: {amount_to_distribute  / (10 ** 12)}")
-
-                async with self.store.lock:
-                    # Get the points of each farmer, as well as payout instructions. Here a chia address is used,
-                    # but other blockchain addresses can also be used.
-                    points_and_ph: List[
-                        Tuple[uint64, bytes]
-                    ] = await self.store.get_farmer_points_and_payout_instructions()
-                    total_points = sum([pt for (pt, ph) in points_and_ph])
-                    if total_points > 0:
-                        mojo_per_point = floor(amount_to_distribute / total_points)
-                        self.log.info(f"Paying out {mojo_per_point} mojo / point")
-
-                        additions_sub_list: List[Dict] = [
-                            {"puzzle_hash": self.pool_fee_puzzle_hash, "amount": pool_coin_amount}
-                        ]
-                        for points, ph in points_and_ph:
-                            if points > 0:
-                                additions_sub_list.append({"puzzle_hash": ph, "amount": points * mojo_per_point})
-
-                            if len(additions_sub_list) == self.max_additions_per_transaction:
-                                await self.pending_payments.put(additions_sub_list.copy())
-                                self.log.info(f"Will make payments: {additions_sub_list}")
-                                additions_sub_list = []
-
-                        if len(additions_sub_list) > 0:
-                            self.log.info(f"Will make payments: {additions_sub_list}")
-                            await self.pending_payments.put(additions_sub_list.copy())
-
-                        # Subtract the points from each farmer
-                        await self.store.clear_farmer_points()
-                    else:
-                        self.log.info(f"No points for any farmer. Waiting {self.payment_interval}")
-
-                await asyncio.sleep(self.payment_interval)
-            except asyncio.CancelledError:
-                self.log.info("Cancelled create_payments_loop, closing")
-                return
-            except Exception as e:
-                error_stack = traceback.format_exc()
-                self.log.error(f"Unexpected error in create_payments_loop: {e} {error_stack}")
-                await asyncio.sleep(self.payment_interval)
-
-    async def submit_payment_loop(self):
-        while True:
-            try:
-                peak_height = self.state_keeper.blockchain_state["peak"].height
-                await self.wallet_rpc_client.log_in_and_skip(fingerprint=self.wallet_fingerprint)
-                if not self.state_keeper.blockchain_state["sync"]["synced"] or not self.state_keeper.wallet_synced:
-                    self.log.warning("Waiting for wallet sync")
-                    await asyncio.sleep(60)
-                    continue
-
-                payment_targets = await self.pending_payments.get()
-                assert len(payment_targets) > 0
-
-                self.log.info(f"Submitting a payment: {payment_targets}")
-
-                # TODO(pool): make sure you have enough to pay the blockchain fee, this will be taken out of the pool
-                # fee itself. Alternatively you can set it to 0 and wait longer
-                # blockchain_fee = 0.00001 * (10 ** 12) * len(payment_targets)
-                blockchain_fee = 0
-                try:
-                    transaction: TransactionRecord = await self.wallet_rpc_client.send_transaction_multi(
-                        self.wallet_id, payment_targets, fee=blockchain_fee
-                    )
-                except ValueError as e:
-                    self.log.error(f"Error making payment: {e}")
-                    await asyncio.sleep(10)
-                    await self.pending_payments.put(payment_targets)
-                    continue
-
-                self.log.info(f"Transaction: {transaction}")
-
-                while (
-                    not transaction.confirmed
-                    or not (peak_height - transaction.confirmed_at_height) > self.confirmation_security_threshold
-                ):
-                    transaction = await self.wallet_rpc_client.get_transaction(self.wallet_id, transaction.name)
-                    peak_height = self.state_keeper.blockchain_state["peak"].height
-                    self.log.info(
-                        f"Waiting for transaction to obtain {self.confirmation_security_threshold} confirmations"
-                    )
-                    if not transaction.confirmed:
-                        self.log.info(f"Not confirmed. In mempool? {transaction.is_in_mempool()}")
-                    else:
-                        self.log.info(f"Confirmations: {peak_height - transaction.confirmed_at_height}")
-                    await asyncio.sleep(10)
-
-                # TODO(pool): persist in DB
-                self.log.info(f"Successfully confirmed payments {payment_targets}")
-
-            except asyncio.CancelledError:
-                self.log.info("Cancelled submit_payment_loop, closing")
-                return
-            except Exception as e:
-                # TODO(pool): retry transaction if failed
-                self.log.error(f"Unexpected error in submit_payment_loop: {e}")
-                await asyncio.sleep(60)
 
     async def confirm_partials_loop(self):
         """
@@ -596,8 +320,8 @@ class Pool:
                 request.payload.payout_instructions,
                 True,
             )
-            self.scan_p2_singleton_puzzle_hashes.add(p2_singleton_puzzle_hash)
             await self.store.add_farmer_record(farmer_record)
+            self.payment_manager.register_new_farmer(farmer_record)
 
             return PostFarmerResponse(self.welcome_message).to_json_dict()
