+++ pool/pool_server.py
@@ -31,6 +31,7 @@ from chia.util.config import load_config
 from .difficulty_adjustment import get_new_difficulty
 from .record import FarmerRecord
 from .pool import Pool
+from .payment_manger.abstract import AbstractPaymentManager
 from .store.abstract import AbstractPoolStore
 from .util import error_response
 
@@ -51,10 +52,11 @@ def check_authentication_token(launcher_id: bytes32, token: uint64, timeout: uin
 
 class PoolServer:
     def __init__(self, config: Dict, constants: ConsensusConstants, pool_store: Optional[AbstractPoolStore] = None,
-                 difficulty_function: Callable = get_new_difficulty):
+                 difficulty_function: Callable = get_new_difficulty,
+                 payment_manager: Optional[AbstractPaymentManager] = None):
 
         self.log = logging.getLogger(__name__)
-        self.pool = Pool(config, constants, pool_store, difficulty_function)
+        self.pool = Pool(config, constants, pool_store, difficulty_function, payment_manager)
 
     async def start(self):
         await self.pool.start()
